+++ inc/how-it-works/streamgraph.php
@@ -7,6 +7,7 @@
 
             <p>
                 <span class="number">1</span>
+                <img class="howimg" src="https://openknowledgemaps.org/img/howitworks/SG-max-1000.png" alt=''>
             </p> 
             <p>A Streamgraph presents you with an overview of the main keywords over time based on up to 1000 documents matching your query.
             </p>
@@ -17,8 +18,9 @@
 
             <p>
                 <span class="number">2</span>
+                <img class="howimg" src="https://openknowledgemaps.org/img/howitworks/SG-stream.png" alt=''>
             </p>
-            <p>The main keywords and related documents are represented as colored streams.
+            <p>The main keywords and related documents are represented as coloured streams.
             </p>
 
         </div>
@@ -27,6 +29,7 @@
 
             <p>
                 <span class="number">3</span>
+                <img class="howimg" src="https://openknowledgemaps.org/img/howitworks/SG-graph.png" alt=''>
             </p>
             <p>Streamgraphs are particularly useful for investigating the evolution of keywords over time and to analyse trends in research. 
             </p>