+++ js/search_form.js
@@ -8,6 +8,7 @@ var SearchOptions = {
     },
     
     drawOptions: function(tag, data) {
+        const self = this;
         data.options.forEach(function(option) {
             let label = d3.select(tag).append("label")
                     .attr("class", "radio-inline")
@@ -16,6 +17,7 @@ var SearchOptions = {
                     .attr("type", "radio")
                     .attr("name", "optradio")
                     .attr("value", option.id)
+                    .on("click", () => self.trackOptionEvent("change", "data_source", option.id));
             
             if(option.default && !option.disabled) {
                 radio_button.attr("checked", true)
@@ -43,6 +45,7 @@ var SearchOptions = {
     },
     
     drawExamples: function(tag, data) {
+        const self = this;
         d3.select(tag).text(data.example_text);
         let examples = d3.select(tag).append("span")
                                 .attr("class", "map-examples")
@@ -52,7 +55,8 @@ var SearchOptions = {
                     .attr("class", "underline")
                     .attr("target", "_blank")
                     .attr("href", example.link)
-                    .text(example.text);
+                    .text(example.text)
+                    .on("click", () => self.trackMatomoEvent("Search box", "Open try-out map", example.link));
         })
     },
     
@@ -145,6 +149,7 @@ var SearchOptions = {
                                     .attr("class", input.class)
                                     .attr("type", "text")
                                     .attr("size", "18")
+                                    .on("change", () => self.trackOptionEvent("change", input.id));
                         })
                     }
                 })
@@ -163,25 +168,24 @@ var SearchOptions = {
                 $("#input-container").css("display", "block");
             }
 
+            self.trackMatomoEvent("Search box", closed ? "Hide options" : "Show options", "Options toggle");
         });
 
     },
-    select_multi: function (dropdown_class, entity, width, data) {
-
-        var self = this;
+    select_multi: function (entityID, entityName, width, data) {
+        const self = this;
+        const dropdown_class = '.dropdown_multi_' + entityID;
+        const is_multiple = !!$(dropdown_class).prop("multiple");
 
          $(function () {
             $(dropdown_class).multiselect({
-                allSelectedText: "All " + entity
-                , nonSelectedText: "No " + entity
-                , nSelectedText: entity
+                allSelectedText: "All " + entityName
+                , nonSelectedText: "No " + entityName
+                , nSelectedText: entityName
                 , buttonWidth: width
                 , maxHeight: 250
                 , includeSelectAllOption: true
-                , numberDisplayed: function () {
-                    let is_multiple = $(dropdown_class).prop("multiple");
-                    return is_multiple ? 0 : 1;
-                }()
+                , numberDisplayed: !is_multiple
                 , onChange: function (element, checked) {
                     if (checked === true) {
                         if(dropdown_class === ".dropdown_multi_time_range" 
@@ -197,6 +201,28 @@ var SearchOptions = {
                             self.setDateRangeFromPreset("#from", "#to", element.val(), data.start_date);
                         }
                     }
+
+                    let identifier = entityID;
+                    if (is_multiple) {
+                        if (checked) {
+                            identifier += "_add";
+                        } else {
+                            identifier += "_remove";
+                        }
+                    }
+                    self.trackOptionEvent("change", identifier, element.val());
+                }
+                , onDropdownShown: function() {
+                    self.trackOptionEvent("show", entityID);
+                }
+                , onDropdownHidden: function() {
+                    self.trackOptionEvent("hide", entityID);
+                }
+                , onSelectAll: function() {
+                    self.trackOptionEvent("select_all", entityID);
+                }
+                , onDeselectAll: function() {
+                    self.trackOptionEvent("deselect_all", entityID);
                 }
             });
 
@@ -308,6 +334,7 @@ var SearchOptions = {
                     var d = i.selectedDay;
                     $(this).datepicker('setDate', new Date(y, m - 1, d));
                 },
+                onSelect: () => self.trackOptionEvent("change", from.replace("#", "")),
                 firstDay: 1
             });
             $(to).datepicker({
@@ -319,6 +346,7 @@ var SearchOptions = {
                     var d = i.selectedDay;
                     $(this).datepicker('setDate', new Date(y, m - 1, d));
                 },
+                onSelect: () => self.trackOptionEvent("change", to.replace("#", "")),
                 firstDay: 1
             });
 
@@ -526,6 +554,24 @@ var SearchOptions = {
             return null // unknown lang code
         }
 
+    },
+    trackMatomoEvent: function(category, action, name, value, dimensions) {
+        // _paq is a global variable defined by the matomo script
+        if (typeof _paq !== "undefined") {
+            _paq.push(["trackEvent", category, action, name, value, dimensions]);
+        }
+        if (window.location.hostname.startsWith("localhost")) {
+            console.log("DEBUG MATOMO:", {category, action, name, value, dimensions});
+        }
+    },
+    trackOptionEvent: function(type, option_id, value) {
+        if (event_type_to_name[type][option_id]) {
+            const [action, defaultName] = event_type_to_name[type][option_id];
+            const name = typeof value !== "undefined" ? value : defaultName;
+            this.trackMatomoEvent("Search box", action, name);
+        } else {
+            console.warn(`The type '${type}', option '${option_id}' couldn't be mapped to an event.`);
+        }
     }
 };
 
@@ -553,4 +599,39 @@ var removeDefault = function(id) {
     })
 }
 
-
+var event_type_to_name = {
+    show: {
+        time_range: ["Show timespan options", "Timespan dropdown"],
+        sorting: ["Show sorting options", "Sorting dropdown"],
+        document_types: ["Show doctypes options", "Doctypes dropdown"],
+        article_types: ["Show doctypes options", "Doctypes dropdown"],
+        min_descsize: ["Show quality options", "Metadata quality dropdown"],
+    },
+    hide: {
+        time_range: ["Hide timespan options", "Timespan dropdown"],
+        sorting: ["Hide sorting options", "Sorting dropdown"],
+        document_types: ["Hide doctypes options", "Doctypes dropdown"],
+        article_types: ["Hide doctypes options", "Doctypes dropdown"],
+        min_descsize: ["Hide quality options", "Metadata quality dropdown"],
+    },
+    change: {
+        time_range: ["Change timespan", "Timespan dropdown"],
+        sorting: ["Change sorting", "Sorting dropdown"],
+        document_types_add: ["Add document type", "Doctypes dropdown"],
+        document_types_remove: ["Remove document type", "Doctypes dropdown"],
+        article_types_add: ["Add document type", "Doctypes dropdown"],
+        article_types_remove: ["Remove document type", "Doctypes dropdown"],
+        min_descsize: ["Change quality", "Metadata quality dropdown"],
+        from: ["Change date", "From input"],
+        to: ["Change date", "To input"],
+        data_source: ["Change data source", "Data source radio"],
+    },
+    select_all: {
+        document_types: ["Select all doctypes", "Doctypes dropdown"],
+        article_types: ["Select all doctypes", "Doctypes dropdown"],
+    },
+    deselect_all: {
+        document_types: ["Deselect all doctypes", "Doctypes dropdown"],
+        article_types: ["Deselect all doctypes", "Doctypes dropdown"],
+    }
+}