+++ js/searchbox/components/BasicOptions.js
@@ -1,0 +1,9 @@
+"use strict";
+
+const e = React.createElement;
+
+const BasicOptions = ({ children }) => {
+  return e("div", { className: "basic_options" }, children);
+};
+
+export default BasicOptions;