+++ CHANGELOG.md
@@ -1,5 +1,53 @@
 # Changelog
 
+## 2022-07-20
+
+### Changes
+
+- Customized UI so that it works with streamgraphs in BASE too.
+
+## 2022-07-14
+
+### Changes
+
+- Improved search box event tracking structure (tracking dropdown options).
+
+## 2022-06-23
+
+### New features
+
+- New frontend component: search box
+- New swagger for setting the new search box parameters
+
+### Changes
+
+- GET endpoint swagger improvements
+- Improved CS fail page links & behavior
+
+## 2022-03-28
+
+### Changes
+
+- Headstart FAQs button is hidden in streamgraphs.
+
+## 2022-03-17
+
+### Changes
+
+- Improved the page metadata export to provide better information to the Zotero add-on.
+
+## 2022-03-03
+
+### Changes
+
+- if no map id is specified, the app doesn't crash
+
+## 2022-02-17
+
+### Removal
+
+- removed citation CSS (moved to Headstart)
+
 ## 2022-01-31
 
 ### New features